V1 updates

+++ src/main/java/com/suyash/shopping/model/Outlet.java
@@ -5,7 +5,10 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import javax.validation.constraints.DecimalMax;
 
+import java.math.BigDecimal;
+import java.text.DecimalFormat;
 import java.util.Set;
 
 import static javax.persistence.GenerationType.IDENTITY;
@@ -21,8 +24,8 @@ public class Outlet {
     private String storeName;
     private String area;
     private String pincode;
-    private Long latitude;
-    private Long longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     @OneToMany(mappedBy = "outlet")
     private Set<Inventory> inventory;
 }
